Fix more build issues

+++ src/chessboard.cpp
@@ -14,7 +14,7 @@ void ChessBoard::clear_board() {
         bitboards[i] = 0;
     }
 
-    for (int i = 0; i < pieces.size(); i++) {
+    for (size_t i = 0; i < pieces.size(); i++) {
         pieces[i] = 0;
     }
 }
